Refactor Admin Page
Update the Admin page to focus solely on the management of APIs and users, removing any unrelated features or components.
[skip gpt_engineer]
