+++ src/components/admin/UserManagement.tsx
@@ -1,0 +1,133 @@
+import { useState } from "react";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Input } from "@/components/ui/input";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { Button } from "@/components/ui/button";
+import { useToast } from "@/hooks/use-toast";
+import { supabase } from "@/integrations/supabase/client";
+import { useQuery } from "@tanstack/react-query";
+import { User } from "@supabase/supabase-js";
+import { Users } from "lucide-react";
+
+export function UserManagement() {
+  const [searchQuery, setSearchQuery] = useState("");
+  const [statusFilter, setStatusFilter] = useState("all");
+  const { toast } = useToast();
+
+  const { data: users, isLoading } = useQuery({
+    queryKey: ['admin-users'],
+    queryFn: async () => {
+      const { data: { users }, error } = await supabase.auth.admin.listUsers();
+      if (error) throw error;
+      return users as User[];
+    },
+  });
+
+  // Filter users based on search and status
+  const filteredUsers = users?.filter(user => {
+    const matchesSearch = user.email?.toLowerCase().includes(searchQuery.toLowerCase());
+    const matchesStatus = statusFilter === "all" || 
+      (statusFilter === "active" && user.last_sign_in_at) ||
+      (statusFilter === "inactive" && !user.last_sign_in_at);
+    return matchesSearch && matchesStatus;
+  });
+
+  const handleDeleteUser = async (userId: string) => {
+    try {
+      const { error } = await supabase.auth.admin.deleteUser(userId);
+      if (error) throw error;
+      
+      toast({
+        title: "User deleted successfully",
+        description: "The user has been removed from the system.",
+      });
+    } catch (error: any) {
+      toast({
+        title: "Error deleting user",
+        description: error.message,
+        variant: "destructive",
+      });
+    }
+  };
+
+  return (
+    <Card className="bg-white border-gray-200">
+      <CardHeader>
+        <CardTitle className="text-black text-2xl font-spotify flex items-center gap-2">
+          <Users className="w-6 h-6" />
+          User Management
+        </CardTitle>
+      </CardHeader>
+      <CardContent>
+        <div className="flex gap-4 mb-6">
+          <Input
+            placeholder="Search users..."
+            value={searchQuery}
+            onChange={(e) => setSearchQuery(e.target.value)}
+            className="bg-white border-gray-200 text-black"
+          />
+          <Select value={statusFilter} onValueChange={setStatusFilter}>
+            <SelectTrigger className="w-[180px] bg-white border-gray-200 text-black">
+              <SelectValue placeholder="Filter by status" />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectItem value="all">All Users</SelectItem>
+              <SelectItem value="active">Active</SelectItem>
+              <SelectItem value="inactive">Inactive</SelectItem>
+            </SelectContent>
+          </Select>
+        </div>
+
+        <div className="overflow-x-auto">
+          <table className="w-full">
+            <thead>
+              <tr className="border-b border-gray-200">
+                <th className="text-left py-3 px-4 text-black font-spotify">Email</th>
+                <th className="text-left py-3 px-4 text-black font-spotify">Status</th>
+                <th className="text-left py-3 px-4 text-black font-spotify">Joined</th>
+                <th className="text-left py-3 px-4 text-black font-spotify">Actions</th>
+              </tr>
+            </thead>
+            <tbody>
+              {isLoading ? (
+                <tr>
+                  <td colSpan={4} className="text-center py-4 text-gray-600">Loading users...</td>
+                </tr>
+              ) : filteredUsers?.map((user) => (
+                <tr key={user.id} className="border-b border-gray-100">
+                  <td className="py-3 px-4 text-black">{user.email}</td>
+                  <td className="py-3 px-4">
+                    <span className={`px-2 py-1 rounded-full text-xs ${
+                      user.last_sign_in_at ? 'bg-green-100 text-green-800' : 'bg-gray-100 text-gray-800'
+                    } font-medium`}>
+                      {user.last_sign_in_at ? 'Active' : 'Inactive'}
+                    </span>
+                  </td>
+                  <td className="py-3 px-4 text-gray-600">
+                    {new Date(user.created_at).toLocaleDateString()}
+                  </td>
+                  <td className="py-3 px-4">
+                    <Button 
+                      variant="ghost" 
+                      size="sm" 
+                      className="text-red-500 hover:text-red-600 hover:bg-red-50"
+                      onClick={() => handleDeleteUser(user.id)}
+                    >
+                      Delete
+                    </Button>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+      </CardContent>
+    </Card>
+  );
+}
